requirements.txt
asgiref==3.8.1
Django==4.2
django-modeltranslation==0.18.11
pillow==11.0.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2



